refactor(compare): extract C3 to checks/missing_values.py (Phase A 6/N)
Move check_c3_missing_accounting (missing-value rate comparison) into
the checks subpackage.

Regression-check: COPDGene, ARIC, HCHS — MD + JSON all match. Tests:
79/79 pass.

Main file: 1584 → 1524 lines.
